+++ src/Paket.Core/PublicAPI.fs
@@ -5,6 +5,7 @@ open Paket.Logging
 open Paket.PackageSources
 
 open System
+open System.Xml
 open System.IO
 open Chessie.ErrorHandling
 open InstallProcess
@@ -667,6 +668,50 @@ type Dependencies(dependenciesFileName: string) =
                          with
                            | _ -> None)
         |> Array.toList
+
+
+    /// Fix the transitive references in a list of generated .nuspec files
+    static member FixNuspecs (referencesFile:string, nuspecFileList:string list) =
+
+        for nuspecFile in nuspecFileList do
+            if not (File.Exists nuspecFile) then
+                failwithf "Specified file '%s' does not exist." nuspecFile
+
+        for nuspecFile in nuspecFileList do
+            let nuspecText = File.ReadAllText nuspecFile
+
+            let doc =
+                try let doc = Xml.XmlDocument() in doc.LoadXml nuspecText
+                    doc
+                with exn -> failwithf "Could not parse nuspec file '%s'.%sMessage: %s" nuspecFile Environment.NewLine exn.Message
+
+            if not (File.Exists referencesFile) then
+                failwithf "Specified references-file '%s' does not exist." referencesFile
+
+            let referencesText = File.ReadAllLines referencesFile
+            let transitiveReferences =
+                referencesText |> Array.map (fun l -> l.Split [|','|])
+                |> Array.choose (fun x -> if x.[2] = "Transitive" then Some x.[0] else None)
+                |> Set.ofArray
+
+            let rec traverse (parent:XmlNode) =
+                let nodesToRemove = ResizeArray()
+                for node in parent.ChildNodes do
+                    if node.Name = "dependency" then
+                        let packageName = 
+                            match node.Attributes.["id"] with null -> "" | x -> x.InnerText
+
+                        if transitiveReferences.Contains packageName then
+                            nodesToRemove.Add node |> ignore
+
+                if nodesToRemove.Count = 0 then
+                    for node in parent.ChildNodes do traverse node
+                else
+                    for node in nodesToRemove do parent.RemoveChild node |> ignore
+            traverse doc
+            use fileStream = File.Open (nuspecFile, FileMode.Create)
+            doc.Save fileStream
+
 
 module PublicAPI =
     /// Takes a version string formatted for Semantic Versioning and parses it